fix gpu type

+++ packages/postgres-database/src/simcore_postgres_database/migration/versions/e0a2557dec27_add_services_limitations.py
@@ -55,7 +55,7 @@ def upgrade():
         sa.Column("ram", sa.BigInteger(), nullable=True),
         sa.Column("cpu", sa.Numeric(), nullable=True),
         sa.Column("vram", sa.BigInteger(), nullable=True),
-        sa.Column("gpu", sa.Numeric(), nullable=True),
+        sa.Column("gpu", sa.Integer(), nullable=True),
         sa.Column(
             "created",
             sa.DateTime(timezone=True),
